+++ mobile/android/android-components/components/feature/customtabs/src/main/java/mozilla/components/feature/customtabs/feature/OriginVerifierFeature.kt
@@ -9,7 +9,6 @@ import android.net.Uri
 import androidx.annotation.VisibleForTesting
 import androidx.browser.customtabs.CustomTabsService.Relation
 import androidx.browser.customtabs.CustomTabsSessionToken
-import mozilla.components.concept.fetch.Client
 import mozilla.components.feature.customtabs.store.CustomTabState
 import mozilla.components.feature.customtabs.store.CustomTabsAction
 import mozilla.components.feature.customtabs.store.OriginRelationPair
@@ -18,13 +17,15 @@ import mozilla.components.feature.customtabs.store.VerificationStatus.FAILURE
 import mozilla.components.feature.customtabs.store.VerificationStatus.PENDING
 import mozilla.components.feature.customtabs.store.VerificationStatus.SUCCESS
 import mozilla.components.feature.customtabs.verify.OriginVerifier
+import mozilla.components.service.digitalassetlinks.RelationChecker
 
 class OriginVerifierFeature(
-    private val httpClient: Client,
     private val packageManager: PackageManager,
-    @VisibleForTesting internal val apiKey: String?,
+    private val relationChecker: RelationChecker,
     private val dispatch: (CustomTabsAction) -> Unit
 ) {
+
+    private var cachedVerifier: Triple<String, Int, OriginVerifier>? = null
 
     suspend fun verify(
         state: CustomTabState,
@@ -51,6 +52,16 @@ class OriginVerifierFeature(
     }
 
     @VisibleForTesting
-    internal fun getVerifier(packageName: String, @Relation relation: Int) =
-        OriginVerifier(packageName, relation, packageManager, httpClient, apiKey)
+    internal fun getVerifier(packageName: String, @Relation relation: Int): OriginVerifier {
+        cachedVerifier?.let {
+            val (cachedPackage, cachedRelation, verifier) = it
+            if (cachedPackage == packageName && cachedRelation == relation) {
+                return verifier
+            }
+        }
+
+        return OriginVerifier(packageName, relation, packageManager, relationChecker).also {
+            cachedVerifier = Triple(packageName, relation, it)
+        }
+    }
 }