+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/DefaultComponents.kt
@@ -38,10 +38,10 @@ import mozilla.components.concept.engine.Engine
 import mozilla.components.concept.fetch.Client
 import mozilla.components.feature.addons.AddonManager
 import mozilla.components.feature.addons.amo.AddonCollectionProvider
-import mozilla.components.feature.addons.update.AddonUpdater
-import mozilla.components.feature.addons.update.DefaultAddonUpdater
 import mozilla.components.feature.addons.migration.DefaultSupportedAddonsChecker
 import mozilla.components.feature.addons.migration.SupportedAddonsChecker
+import mozilla.components.feature.addons.update.AddonUpdater
+import mozilla.components.feature.addons.update.DefaultAddonUpdater
 import mozilla.components.feature.app.links.AppLinksInterceptor
 import mozilla.components.feature.app.links.AppLinksUseCases
 import mozilla.components.feature.contextmenu.ContextMenuUseCases
@@ -65,12 +65,14 @@ import mozilla.components.feature.tabs.TabsUseCases
 import mozilla.components.feature.webnotifications.WebNotificationFeature
 import mozilla.components.lib.fetch.httpurlconnection.HttpURLConnectionClient
 import mozilla.components.lib.nearby.NearbyConnection
+import mozilla.components.service.digitalassetlinks.local.StatementApi
+import mozilla.components.service.digitalassetlinks.local.StatementRelationChecker
 import org.mozilla.samples.browser.addons.AddonsActivity
 import org.mozilla.samples.browser.downloads.DownloadService
 import org.mozilla.samples.browser.ext.components
 import org.mozilla.samples.browser.integration.FindInPageIntegration
-import org.mozilla.samples.browser.media.MediaService
 import org.mozilla.samples.browser.integration.P2PIntegration
+import org.mozilla.samples.browser.media.MediaService
 import org.mozilla.samples.browser.request.SampleRequestInterceptor
 import java.util.concurrent.TimeUnit
 
@@ -203,6 +205,11 @@ open class DefaultComponents(private val applicationContext: Context) {
         NearbyConnection(applicationContext)
     }
 
+    // Digital Asset Links checking
+    val relationChecker by lazy {
+        StatementRelationChecker(StatementApi(client))
+    }
+
     // Intent
     val tabIntentProcessor by lazy {
         TabIntentProcessor(sessionManager, sessionUseCases.loadUrl, searchUseCases.newTabSearch)
@@ -213,9 +220,8 @@ open class DefaultComponents(private val applicationContext: Context) {
             TrustedWebActivityIntentProcessor(
                 sessionManager,
                 sessionUseCases.loadUrl,
-                client,
                 applicationContext.packageManager,
-                null,
+                relationChecker,
                 customTabsStore
             ),
             CustomTabIntentProcessor(sessionManager, sessionUseCases.loadUrl, applicationContext.resources)